+++ lib/Consts/Consts.h
@@ -10,25 +10,25 @@ enum STATUS {
   CLEARING        //Transaction finished, look for negative container weight then tare
 };
 
-const int BUTTON_PIN = 5;     // the number of the pushbutton pin
-const int GREEN_LED_PIN = 6;
-const int RED_LED_PIN = 7;
-const int POT_PIN = A0;    // select the input pin for the potentiometer
-const int LOOP_DELTA = 20;
-const int BASE_CALIBRATION_OFFSET = 200;
-const int PRICE_PER_CUP = 50; //cents per 250ml
+const uint8_t BUTTON_PIN = 5;     // the number of the pushbutton pin
+const uint8_t GREEN_LED_PIN = 6;
+const uint8_t RED_LED_PIN = 7;
+const uint8_t POT_PIN = A0;    // select the input pin for the potentiometer
+const uint8_t LOOP_DELTA = 20;
+const uint8_t BASE_CALIBRATION_OFFSET = 200;
+const uint8_t PRICE_PER_CUP = 50; //cents per 250ml
 
 // LCD Pins
-const int OLED_RESET = 4;
+const uint8_t OLED_RESET = 4;
 
 // Scale pins
-const int DOUT = 3;
-const int CLK = 2;
+const uint8_t DOUT = 3;
+const uint8_t CLK = 2;
 
 // For RFID Reader
-const int RST_PIN = 9;
-const int SS_PIN = 53;
+const uint8_t RST_PIN = 9;
+const uint8_t SS_PIN = 53;
 
-String enumToString(int sys);
+String enumToString(uint8_t sys);
 
 #endif
